Create packages and move java files in corresponding packages.
